refactor: Move withLocales inside components
We must not wrap MobileRouter in the I18n otherwise the t from the app
is overriden. This is why it is done in "children" components like
Authentication and Revoked
