+++ packages/cozy-authentication/src/withLocales.jsx
@@ -1,0 +1,17 @@
+import { I18n } from 'cozy-ui/transpiled/react/I18n'
+
+const locales = {
+  en: require(`./locales/en.json`),
+  fr: require(`./locales/fr.json`)
+}
+
+const withLocales = Component =>
+  translate()(props => {
+    return (
+      <I18n dictRequire={localeCode => locales[localeCode]} lang={props.lang}>
+        <Component {...props} />
+      </I18n>
+    )
+  })
+
+export default withLocales
